add auth

+++ server/src/controllers/users.js
@@ -4,12 +4,19 @@ const jwt = require('jsonwebtoken')
 
 const databaseConnect = require('../../database-connecting')
 const { token_key } = require('../../environment-configs')
+const e = require('express')
 const tableName = 'users_table'
 
 module.exports = {
   getUsers: async (req, res, next) => {
+    const { user_role, user_id } = req.user
     try {
-      const response = await databaseConnect(tableName).select()
+      let response
+      if (user_role === 'admin') {
+        response = await databaseConnect(tableName).select()
+      } else {
+        response = await databaseConnect(tableName).where('id', user_id)
+      }
       res.status(stausCode.OK).send({ results: response })
     } catch (error) {
       next(error)
@@ -25,46 +32,49 @@ module.exports = {
         const token = jwt.sign(
           {
             user_id: response[0].id,
+            user_role: response[0].role,
+            name: response[0].name,
           },
           token_key,
           {
             expiresIn: '2h',
           },
         )
-        response.push({ token: token })
+        response[0].token = token
         res.status(stausCode.OK).send(response)
+      } else {
+        res.status(stausCode.BAD_REQUEST).send({ message: 'Invaild Credentials' })
       }
-      res.status(stausCode.BAD_REQUEST).send({ message: 'Invaild Credentials' })
     } catch (error) {
       next(error)
     }
   },
 
   register: async (req, res, next) => {
-    const { username, password, name } = req.body
+    const { username, password, name, role = 'user' } = req.body
     try {
       encryptedPassword = await bcrypy.hash(password, 10)
       const response = await databaseConnect(tableName).insert({
         username: username.toLowerCase(),
         password: encryptedPassword,
         name: name,
+        role: role,
       })
       res.status(stausCode.CREATED).send({ results: response })
     } catch (error) {
       error.status = 409
+      error.message = 'the user exist'
       next(error)
     }
   },
 
   edit: async (req, res, next) => {
-    const { id } = req.params
+    const { user_id } = req.user
     const { name, password } = req.body
-
-    console.log(id, name, password)
     try {
       encryptedPassword = await bcrypy.hash(password, 10)
       const response = await databaseConnect(tableName)
-        .where('id', id)
+        .where('id', user_id)
         .update({ name: name, password: encryptedPassword })
       res.status(stausCode.OK).send({ results: response })
     } catch (error) {
@@ -74,9 +84,14 @@ module.exports = {
 
   delete: async (req, res, next) => {
     const { id } = req.params
+    const { user_role, user_id } = req.user
     try {
-      const response = await databaseConnect(tableName).where('id', id).delete()
-      res.status(stausCode.OK).send({ message: 'Deleted', response })
+      if (user_id != id && user_role === 'admin') {
+        const response = await databaseConnect(tableName).where('id', id).delete()
+        res.status(stausCode.OK).send({ message: 'Deleted', response })
+      } else {
+        res.status(stausCode.FORBIDDEN).send({ message: 'Permission denied' })
+      }
     } catch (error) {
       next(error)
     }
